% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deposition.r
\name{deposition.default}
\alias{deposition.default}
\title{Simulate the carbon balance of a soil profile experiencing deposition of soil transported to the depositional location from an eroding profile}
\usage{
deposition.default(dep_profile, ero_process, time,
  k_ero = ero_process$Profile$Profile$k, I_ero = NULL, n_ox_ero,
  n_prod_ero, k_dep = dep_profile$Profile$k, I_dep = NULL, n_ox_dep,
  n_prod_dep, frac_oxidized = rep(0, length(ero_process[[1]])))
}
\arguments{
\item{dep_profile}{A profile data frame object generated by SOrCERO::gen_profile_SOC().}

\item{ero_process}{An erosional profile list object generated by SOrCERO::erosion().}

\item{time}{A number representing the amount of time steps to simulate.}

\item{k_ero}{A numeric vector representing the natural log of the oxidation first order rate constant, ln(-k), for each layer of the eroded profile.}

\item{I_ero}{A numeric vector representing the SOC production, I, in units of kg/m^2/time for each layer of the eroded profile.}

\item{n_ox_ero}{A numeric vector representing the oxidation mixing coefficient for each layer of the eroded profile.}

\item{n_prod_ero}{A numeric vector representing the production mixing coefficient for each layer of the eroded profile.}

\item{k_dep}{A numeric vector representing the natural log of the oxidation first order rate constant, ln(-k), for each layer of the depositional profile.}

\item{I_dep}{A numeric vector representing the SOC production, I, in units of kg/m^2/time for each layer of the depositional profile.}

\item{n_ox_dep}{A numeric vector representing the oxidation mixing coefficient for each layer of the depositional profile.}

\item{n_prod_dep}{A numeric vector representing the production mixing coefficient for each layer of the depositional profile.}

\item{frac_oxidized}{A numeric vector representing the fraction of eroded SOC which is oxidized in transport. Default is 0 for the entire profile. Minimum length of frac_oxidized is equal to the time argument.}
}
\value{
A list of two parts. The first, "Depositional Process" is another list whose entries represent the depositional profile at a given time step, starting with 0. The second object in the list, "Dep Timestep Metrics", gives a dataframe of useful metrics with rows as time step.
}
\description{
Simulate the carbon balance of a soil profile experiencing deposition of soil transported to the depositional location from an eroding profile
}
