% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erosion.r
\name{erosion}
\alias{erosion}
\title{Simulate an eroding soil profile's soil carbon balance}
\usage{
erosion(profile, time, k = profile$Profile$k, I = NULL, n_ox, n_prod)
}
\arguments{
\item{profile}{An object of class "SoilProfile" object generated by SOrCERO::gen_profile_SOC().}

\item{time}{A number representing the amount of time steps to simulate.}

\item{k}{A numeric vector representing the natural log of the oxidation first order rate constant, ln(-k), for each layer of the eroded profile.}

\item{I}{A numeric vector representing the SOC production, I, in units of kg/m^2/time for each layer of the eroded profile.}

\item{n_ox}{A numeric vector representing the oxidation mixing coefficient for each layer of the eroded profile.}

\item{n_prod}{A numeric vector representing the production mixing coefficient for each layer of the eroded profile.}
}
\value{
A list of three parts. The first, "Erosion Process" is another list whose entries represent the erosional profile at a given timestep, starting with 0. The second object in the list, "Timestep Metrics", gives a dataframe of useful metrics with rows as time step. The third object holds the original soil profile object.
}
\description{
Simulate an eroding soil profile's soil carbon balance
}
