% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_profile_soc.R
\name{gen_profile_SOC}
\alias{gen_profile_SOC}
\title{Generate a soil profile for use in an erosional or depositional site}
\usage{
gen_profile_SOC(layers, SOC, bulk_density, weight_fraction_c, layer_range,
  erosion_rate_mt, min_age, max_age, interpolate = c(TRUE, FALSE),
  calculate = c(TRUE, FALSE), ero_accel = rep(1, length(layers)))
}
\arguments{
\item{layers}{A numeric vector of soil layers, not measured in units of depth.}

\item{SOC}{An optional numeric vector of SOC at depths corresponding to layers.}

\item{bulk_density}{A numeric vector of soil bulk density at depths corresponding to layers. Optional if SOC is present.}

\item{weight_fraction_c}{A numeric vector of the fraction of soil weight that carbon represents at depths corresponding to layers. Optional if SOC is present.}

\item{layer_range}{A numeric vector of the range of layers desired in the output. Defaults to 1:length(layers)}

\item{erosion_rate_mt}{A scalar indicating the erosion rate in units of meters/time.}

\item{interpolate}{A boolean indicating whether the user would like gen_profile_SOC to interpolate SOC values out to the depth given in layer_range.}

\item{calculate}{A boolean indiating whether the user would like gen_profile_SOC to calculate SOC from weight_fraction_c and bulk_density. Should be FALSE if SOC is supplied.}

\item{ero_accel}{A numeric vector which gives the factor by which the erosion rate should be scaled by for each time step.}
}
\value{
A data frame of n rows and 6 columns where n is equal to layer_range and represents the number of layers in the profile to be eroded. Columns are layers, SOC, depth, k, weight fraction of C, and bulk density
}
\description{
Generate a soil profile for use in an erosional or depositional site
}
